requirements.txt
autopep8==1.4.3
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
coveralls==1.5.1
Django==2.1.2
django-cors-middleware==1.2.0
django-extensions==1.7.1
django-filter==2.0.0
django-jet==1.0.8
django-rest-swagger==2.2.0
django-simple-history==2.7.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
factory-boy==2.11.1
Faker==0.9.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
pkg-resources==0.0.0
pycodestyle==2.5.0
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.5
requests==2.21.0
simplejson==3.16.0
six==1.10.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.23
